feat: seperate metrics card components
